added char validation in student personnel and visitor

+++ MainSystem/edit_info_personnel.py
@@ -412,26 +412,44 @@ class EditPersonnelApp:
                         personnel_contact_num_var.startswith("-")
                         and personnel_contact_num_var[1:].isdigit()
                     ):
-                        self.sql_query.update_personnel(
-                            personnel_num_var,
-                            personnel_firstname_var,
-                            personnel_lastname_var,
-                            personnel_middlename_var,
-                            personnel_contact_num_var,
-                            personnel_address_var,
-                            personnel_type_variable,
-                            pesonnel_status_var,
-                        )
-                        if os.path.exists(self.img_path + "/000000000.jpg"):
-                            img_name = self.personnel_num
-                            os.rename(
-                                self.img_path + "/" + img_name + ".jpg",
-                                self.img_path + "/000000000.jpg",
+                        if personnel_firstname_var.replace(" ", "").isalpha():
+                            if personnel_lastname_var.replace(" ", "").isalpha():
+                                if personnel_middlename_var.replace(" ", "").isalpha():
+                                    self.sql_query.update_personnel(
+                                        personnel_num_var,
+                                        personnel_firstname_var,
+                                        personnel_lastname_var,
+                                        personnel_middlename_var,
+                                        personnel_contact_num_var,
+                                        personnel_address_var,
+                                        personnel_type_variable,
+                                        pesonnel_status_var,
+                                    )
+                                    if os.path.exists(self.img_path + "/000000000.jpg"):
+                                        img_name = self.personnel_num
+                                        os.rename(
+                                            self.img_path + "/" + img_name + ".jpg",
+                                            self.img_path + "/000000000.jpg",
+                                        )
+                                    messbx.showinfo(
+                                        "Success",
+                                        "The personnel record has been edited successfully.",
+                                    )
+                                else:
+                                    messbx.showwarning(
+                                        "Warning",
+                                        "The input for the firstname or lastname or middlename is not a valid character.",
+                                    )
+                            else:
+                                messbx.showwarning(
+                                    "Warning",
+                                    "The input for the firstname or lastname or middlename is not a valid character.",
+                                )
+                        else:
+                            messbx.showwarning(
+                                "Warning",
+                                "The input for the firstname or lastname or middlename is not a valid character.",
                             )
-                        messbx.showinfo(
-                            "Success",
-                            "The personnel record has been edited successfully.",
-                        )
                     else:
                         messbx.showwarning(
                             "Warning",
