+++ MainSystem/register_personnel.py
@@ -278,24 +278,43 @@ class RegisterPersonnelApp:
                         contact_num_var.startswith("-")
                         and contact_num_var[1:].isdigit()
                     ):
-                        self.sql_query.register_personnel(
-                            personnel_num_var,
-                            first_name_var,
-                            last_name_var,
-                            mid_name_var,
-                            contact_num_var,
-                            address_var,
-                            personnel_type_var,
-                        )
-                        img_name = personnel_num_var
-                        os.rename(
-                            self.img_path + "/000000000.jpg",
-                            self.img_path + "/" + img_name + ".jpg",
-                        )
-                        messbx.showinfo(
-                            "Success",
-                            "The personnel record has been registered successfully.",
-                        )
+                        if first_name_var.replace(" ", "").isalpha():
+                            if last_name_var.replace(" ", "").isalpha():
+                                if mid_name_var.replace(" ", "").isalpha():
+
+                                    self.sql_query.register_personnel(
+                                        personnel_num_var,
+                                        first_name_var,
+                                        last_name_var,
+                                        mid_name_var,
+                                        contact_num_var,
+                                        address_var,
+                                        personnel_type_var,
+                                    )
+                                    img_name = personnel_num_var
+                                    os.rename(
+                                        self.img_path + "/000000000.jpg",
+                                        self.img_path + "/" + img_name + ".jpg",
+                                    )
+                                    messbx.showinfo(
+                                        "Success",
+                                        "The personnel record has been registered successfully.",
+                                    )
+                                else:
+                                    messbx.showwarning(
+                                        "Warning",
+                                        "The input for the firstname or lastname or middlename is not a valid character.",
+                                    )
+                            else:
+                                messbx.showwarning(
+                                    "Warning",
+                                    "The input for the firstname or lastname or middlename is not a valid character.",
+                                )
+                        else:
+                            messbx.showwarning(
+                                "Warning",
+                                "The input for the firstname or lastname or middlename is not a valid character.",
+                            )
                     else:
                         messbx.showwarning(
                             "Warning",
