+++ MainSystem/register_student.py
@@ -277,25 +277,43 @@ class RegisterStudentApp:
                         contact_num_var.startswith("-")
                         and contact_num_var[1:].isdigit()
                     ):
-                        if self.register == True:
-                            self.sql_query.register_student(
-                                student_num_var,
-                                first_name_var,
-                                last_name_var,
-                                mid_name_var,
-                                program_var,
-                                section_var,
-                                contact_num_var,
-                                address_var,
-                            )
-                            img_name = student_num_var
-                            os.rename(
-                                self.img_path + "/000000000.jpg",
-                                self.img_path + "/" + img_name + ".jpg",
-                            )
-                            messbx.showinfo(
-                                "Success",
-                                "The student record has been registered successfully.",
+                        if first_name_var.replace(" ", "").isalpha():
+                            if last_name_var.replace(" ", "").isalpha():
+                                if mid_name_var.replace(" ", "").isalpha():
+                                    if self.register == True:
+                                        self.sql_query.register_student(
+                                            student_num_var,
+                                            first_name_var,
+                                            last_name_var,
+                                            mid_name_var,
+                                            program_var,
+                                            section_var,
+                                            contact_num_var,
+                                            address_var,
+                                        )
+                                        img_name = student_num_var
+                                        os.rename(
+                                            self.img_path + "/000000000.jpg",
+                                            self.img_path + "/" + img_name + ".jpg",
+                                        )
+                                        messbx.showinfo(
+                                            "Success",
+                                            "The student record has been registered successfully.",
+                                        )
+                                else:
+                                    messbx.showwarning(
+                                        "Warning",
+                                        "The input for the firstname or lastname or middlename is not a valid character.",
+                                    )
+                            else:
+                                messbx.showwarning(
+                                    "Warning",
+                                    "The input for the firstname or lastname or middlename is not a valid character.",
+                                )
+                        else:
+                            messbx.showwarning(
+                                "Warning",
+                                "The input for the firstname or lastname or middlename is not a valid character.",
                             )
                     else:
                         messbx.showwarning(
